+++ Coosu.Storyboard.Extensions/Computing/EventExtensions.cs
@@ -8,40 +8,40 @@ namespace Coosu.Storyboard.Extensions.Computing
 {
     public static class EventExtensions
     {
-        internal static readonly ReadOnlyDictionary<string, float[]> IneffectiveDictionary = new(
-            new Dictionary<string, float[]>
+        internal static readonly ReadOnlyDictionary<string, double[]> IneffectiveDictionary = new(
+            new Dictionary<string, double[]>
             {
-                [EventTypes.Fade.Flag] = new[] { 0f },
-                [EventTypes.Scale.Flag] = new[] { 0f },
-                [EventTypes.Vector.Flag] = new[] { 0f, 0f },
-                [EventTypes.Color.Flag] = new[] { 0f, 0f, 0f },
+                [EventTypes.Fade.Flag] = new[] { 0d },
+                [EventTypes.Scale.Flag] = new[] { 0d },
+                [EventTypes.Vector.Flag] = new[] { 0d, 0d },
+                [EventTypes.Color.Flag] = new[] { 0d, 0d, 0d },
             });
 
-        internal static readonly ReadOnlyDictionary<string, float[]> DefaultDictionary = new(
-            new Dictionary<string, float[]>
+        internal static readonly ReadOnlyDictionary<string, double[]> DefaultDictionary = new(
+            new Dictionary<string, double[]>
             {
-                [EventTypes.Fade.Flag] = new[] { 1f },
-                [EventTypes.Scale.Flag] = new[] { 1f },
-                [EventTypes.Vector.Flag] = new[] { 1f, 1f },
-                [EventTypes.Rotate.Flag] = new[] { 0f },
-                [EventTypes.Color.Flag] = new[] { 255f, 255f, 255f },
+                [EventTypes.Fade.Flag] = new[] { 1d },
+                [EventTypes.Scale.Flag] = new[] { 1d },
+                [EventTypes.Vector.Flag] = new[] { 1d, 1d },
+                [EventTypes.Rotate.Flag] = new[] { 0d },
+                [EventTypes.Color.Flag] = new[] { 255d, 255d, 255d },
             });
 
-        public static float[] GetIneffectiveValue(this IKeyEvent e)
+        public static double[] GetIneffectiveValue(this IKeyEvent e)
         {
             return IneffectiveDictionary.ContainsKey(e.EventType.Flag)
                 ? IneffectiveDictionary[e.EventType.Flag]
-                : EmptyArray<float>.Value;
+                : EmptyArray<double>.Value;
         }
 
-        public static float[]? GetDefaultValue(this EventType eventType)
+        public static double[]? GetDefaultValue(this EventType eventType)
         {
             return DefaultDictionary.ContainsKey(eventType.Flag)
                 ? DefaultDictionary[eventType.Flag]
                 : null;
         }
 
-        public static float[]? GetDefaultValue(this EventType eventType, ICameraUsable? sprite)
+        public static double[]? GetDefaultValue(this EventType eventType, ICameraUsable? sprite)
         {
             if (sprite == null) return GetDefaultValue(eventType);
 
@@ -51,14 +51,14 @@ namespace Coosu.Storyboard.Extensions.Computing
             return GetDefaultValue(eventType);
         }
 
-        public static List<float> ComputeRelative(this EventType eventType,
-            IReadOnlyList<float> source,
-            IReadOnlyList<float> relativeVal,
+        public static List<double> ComputeRelative(this EventType eventType,
+            IReadOnlyList<double> source,
+            IReadOnlyList<double> relativeVal,
             int? accuracy = null)
         {
             if (eventType.Size < 1)
                 throw new ArgumentOutOfRangeException(nameof(eventType), eventType, "Only support sized event type.");
-            var list = new List<float>(eventType.Size);
+            var list = new List<double>(eventType.Size);
             for (int i = 0; i < eventType.Size; i++)
             {
                 //if (eventType == EventTypes.Fade ||eventType==EventTypes.Scale||eventType==)
@@ -68,7 +68,7 @@ namespace Coosu.Storyboard.Extensions.Computing
                 if (accuracy == null)
                     list.Add(source[i] + relativeVal[i]);
                 else
-                    list.Add((float)Math.Round(source[i] + relativeVal[i], accuracy.Value));
+                    list.Add((double)Math.Round(source[i] + relativeVal[i], accuracy.Value));
             }
 
             return list;
